Load adapters dynamically from a directory.
Any adapters found in the src/adapters/ directory which have a
package.json file will be considered. All adapters are disabled
by default, and are only loaded if they are explicitly enabled in
the settings database.

Zigbee and Z-Wave are temporarily enabled by default, but that will
be backed out in the near future.