Fix develco regressions (#51)
* Keep the cached 16-bit address

for battery powered devices, sometimes the first
thing we hear from them is a broadcast (i.e. addr64 == all F's)
Since we're now doing a network address request for all of the
powered devices, this improves the likelyhood that we'll
recognize which node a first message comes from.

* Bump version to 0.5.6

* Cleanup

Removed clearQueryingSimpleDescriptor function and just clear
the boolean directly.
Noticed a call to parseInt that wasn't passing in the correct base.

* Fix regression for develco sensors

* Use the correct endpoint when reading classifier
attributes (affects bulbs and zoneType)
* Don't issue a read for zoneType if there is a
read already in progress.
* Grab the zoneType from the enrollReq so we don't have
to issue a read for it if we get the enrollReq

* Set the checkin interval as soon as we know the genPollCtrl endpoint

* Also don't write the checkin interval if there is
already a write in progress

* Call populateNodeInfo in more places

Take advantage of more places that battery powered devices
are talking to us and have their receivers enabled.