Add support for deCONZ ConBee
This factors the xbee specifi knowledge out of the existing stack.
zb-driver.js now contains common code for the driver level and
deconz-drive.js and xbee-driver.js contain the zigbee dongle
specific code.

This code also relies on deconz-api, a new repository created for
the deConz dongle which offers a similar API level to the xbee-api.

The ZDO code was also factored out into a separate repository.