Fix calculation for Hue/Saturation
The Color class epxetcs Hue in the range 0-360 and Saturation 0-100.
Zigbee uses 0-254 for both of these.

The code which sets the color:
https://github.com/mozilla-iot/zigbee-adapter/blob/c3a9f90e746e35f6c8c5ef30c111bc53859bf9a2/zb-property.js#L723
does the correct scaling.

This PR fixes the parsing code which takes the zigbee values and converts
them back into RGB colors to apply the correct inverse scaling.