Support garden light (#79)
* Support the Sylvania RGB Garden Light

This cleaned up support for color lights.
The Sylvania RGB Garden Light is the first light that's not
a ZLL light I've encountered, and that wound up uncovering
a coule bugs. This PR should also make the IKEA Trafri outlet
be reported as an outlet. This outlet reports itself as
supporting ZLL, which also uncovered a bug.

* Fixed bug in dumping zcl writeRsp