Improve permit-join and discover
This speeds up the discover process (which is really
a debug feature). Permit join was generating an error
so this just sends the permit join to the diongle
rather than trying to broadcast. I'd like to figure
out how to properly do a broadcast so that all
of the routers are updated about permit-join as well.