Improvements (#104)
* Improve logging for ConBee

When just logging frames, only log the actual
zigbee frames, and not the comms frames between
the host and the dongle

* Improve updating of 16-bit address

* Limit retries for some messages

In particular, during a discover (which is a manual command)
only send a read request out once.
During the scanning, only send out the NETWORK_ADDRESS request
once and only wait a second for the response. This speeds up
the scanning process when there are powered off devices.

* discover improvements

Print a bit more relevant information when doing
a discover.

* Don't try to rebind if we don't know the 16-bit address

* Update dependencies

* Use deviceId rather then colorMode

It turns out that colorMode changes at runtime
to reflect the current mode (i.e. last type of
command sent) rather than a cpability, so it's
unreliable to use it for determining the type of
light.

* Update deconz frame algorithim.

This update the deconz driver to wait for a response
to each request sent to the dongle before sending the next
request. This is not to be confused with sending/receiving zigbee
frames, which happens at a higher layer. This just deals with
communications with with dongle itself.

* Add constants for APS status codes.

* Improve permit-join and discover

This speeds up the discover process (which is really
a debug feature). Permit join was generating an error
so this just sends the permit join to the diongle
rather than trying to broadcast. I'd like to figure
out how to properly do a broadcast so that all
of the routers are updated about permit-join as well.

* Add an extra log so that packets don't fall thry the cracks

* More light related classification improvements.

* Report serial port open errors via errorCallback

* Made corresponding change in xbee driver

permitJoin needs the 64-bit network address

* Address review comments