Added support for the ConBee watchdog timer
With the Conbee the watchdog triggering was largely
invisible since there was a usb-to-serial converter attached
to the AVR, and the AVR reset would only be noticed if there
were commands being processed at the time.

With the ConBee II, the MCU is directly connected to USB, so
when the MCU resets, it would dsiconnect from USB and would
need to be re-enumerated before comms would work.

Kicking the watchdog prevents the dongle from resetting.