Xiaomi Zigbee - Pressure, Motion and illuminance improvements (#141)
WSDCGQ11LM (square) changed the pressure part; unit is now hPa with a minimum of 800 and maximum of 1100 and I used ParseNumericAttr for the parseValueFromAttr. Under normal circumstances the atmospheric pressure falls between 800 and 1100 hPa. I have compared the numbers from the pressure sensor with a nearby weather station and the numbers are comparable.

RTCGQ11LM changed the illuminance part; type is now levelproperty, lux has now a maximum of 1500 and used ParseNumericAttr for the parseValueFromAttr. it is still not very accurate, but it is enough to detect changes in light to act accordingly to it.

Change label from the motion sensors properties from Occupied to Motion.